chg sizes and indices of arrays to be unsigned
git-svn-id: http://vtp.googlecode.com/svn/trunk@2533 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtlib/vtosg/NodeOSG.cpp
@@ -634,7 +634,7 @@ vtMaterial *vtGeom::GetMaterial(int idx)
 {
 	if (!m_pMaterialArray.valid())
 		return NULL;
-	if (idx < 0 || idx >= m_pMaterialArray->GetSize())
+	if (idx < 0 || idx >= (int) m_pMaterialArray->GetSize())
 		return NULL;
 	return m_pMaterialArray->GetAt(idx);
 }
